move: Token 파일 경로 이동
