+++ src/main/java/gift/domain/user/service/UserService.java
@@ -1,6 +1,6 @@
 package gift.domain.user.service;
 
-import gift.auth.jwt.Token;
+import gift.auth.dto.Token;
 import gift.domain.user.repository.UserJpaRepository;
 import gift.domain.user.dto.UserDto;
 import gift.domain.user.dto.UserLoginDto;
